rename default parameters. Delete (fixed) flag
