+++ src/front/shared/helpers/constants/DEFAULT_CURRENCY_PARAMETERS.ts
