+++ src/front/shared/helpers/constants/index.ts
@@ -5,7 +5,7 @@ import modals from './modals'
 import tradeTicker from './TRADE_TICKERS'
 import tokenDecimals from './TOKEN_DECIMALS'
 import pairTypes from './PAIR_TYPES'
-import defaultFeeRates from './DEFAULT_FEE_RATES'
+import defaultCurrencyParameters from './DEFAULT_CURRENCY_PARAMETERS'
 import minAmount from './minAmount'
 import minAmountOffer from './minAmountOffer'
 import coinsWithDynamicFee from './coinsWithDynamicFee'
@@ -19,7 +19,7 @@ export default {
   tokenDecimals,
   notifications,
   pairTypes,
-  defaultFeeRates,
+  defaultCurrencyParameters,
   modals,
   minAmount,
   minAmountOffer,
